move UnifiWifiImage class to image.py
